airflow/requirements-offline.txt
# ============================================================================
# Python Requirements for Offline Airflow Deployment
# ============================================================================
#
# 이 파일은 폐쇄망 환경에서 Airflow를 실행하기 위한 모든 Python 의존성을 정의합니다.
# 버전을 고정하여 재현 가능한 빌드를 보장합니다.
#
# 업데이트 방법:
#   pip install pip-tools
#   pip-compile requirements-offline.in --output-file requirements-offline.txt
# ============================================================================

# ============================================================================
# Database Drivers
# ============================================================================
psycopg2-binary==2.9.9         # PostgreSQL driver
oracledb==2.0.1                # Oracle driver (thin mode, no client needed)

# ============================================================================
# Data Processing
# ============================================================================
pandas==2.1.4
numpy==1.26.2
pandera==0.18.0
pydantic==2.5.3

# ============================================================================
# Search and Vector Database
# ============================================================================
opensearch-py==2.4.2

# ============================================================================
# HTTP Clients
# ============================================================================
httpx==0.25.2
requests==2.31.0

# ============================================================================
# Machine Learning & Embeddings
# ============================================================================
# SentenceTransformers와 모든 의존성
sentence-transformers==2.2.2
transformers==4.36.2
tokenizers==0.15.0
huggingface-hub==0.20.1

# PyTorch CPU-only (용량 절약)
torch==2.1.2+cpu --extra-index-url https://download.pytorch.org/whl/cpu
torchvision==0.16.2+cpu --extra-index-url https://download.pytorch.org/whl/cpu

# ============================================================================
# Additional ML Dependencies
# ============================================================================
scikit-learn==1.3.2
scipy==1.11.4
sympy==1.12
mpmath==1.3.0
pillow==10.1.0
tqdm==4.66.1
regex==2023.12.25
filelock==3.13.1
safetensors==0.4.1

# ============================================================================
# Utility Libraries
# ============================================================================
python-dateutil==2.8.2
pytz==2023.3.post1
typing-extensions==4.9.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
urllib3==2.1.0

# ============================================================================
# YAML/JSON Processing
# ============================================================================
PyYAML==6.0.1

# ============================================================================
# Packaging
# ============================================================================
packaging==23.2
